Move some utility methods to CustomNameUtil, other small fixes
