+++ common/src/main/java/xyz/eclipseisoffline/eclipsescustomname/ParsedPlayerName.java
@@ -7,7 +7,7 @@ import net.minecraft.network.chat.Component;
 public record ParsedPlayerName(String raw, Component parsed) {
     public static final Codec<ParsedPlayerName> CODEC = Codec.STRING.comapFlatMap(raw -> {
         try {
-            return DataResult.success(new ParsedPlayerName(raw, CustomName.argumentToComponent(raw, true, false, false)));
+            return DataResult.success(new ParsedPlayerName(raw, CustomNameUtil.nameArgumentToComponent(raw, true, false, false)));
         } catch (IllegalArgumentException exception) {
             return DataResult.error(() -> "Failed to parse player name: " + exception.getMessage());
         }
